requirements.txt
aws-cdk-lib==2.127
aws-cdk.lambda-layer-kubectl-v29
constructs>=10.0.0,<11.0.0
cdk-bootstrapless-synthesizer